create class Game in lib
